Fix escaping issues in the JSON properly
The problem is that we generate JSON and embed it in javascript in a
single quoted string. The JSON itself uses double quoted strings for the
content and takes care of escaping double quotes no its own, so we're
safe there, but if the content contains a single quote is has to be
escaped.

The previous fix replaced the single ticks with their HTML entity
equivalent which fixes the javascript error but also alters the content
of the JSON. That can lead to problems in the future when we want to
generate download links for the extracted files, for example. This
commit instead simply escapes the single ticks (' -> \') which is what
we really want here.

Example of the resulting file object in the browser before:

  file: Object {name: "/etc/magicapp&#39;.conf", type: "file"...}

now:

  file: Object {name: "/etc/magicapp'.conf", type: "file"...}